User authorization

+++ app/Http/Controllers/ListingController.php
@@ -78,6 +78,11 @@ class ListingController extends Controller
     //Update listing data
     public function update(Request $request, Listing $listing)
     {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -101,6 +106,11 @@ class ListingController extends Controller
     //Delete listing
     public function destroy(Listing $listing)
     {
+        //Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
@@ -108,6 +118,6 @@ class ListingController extends Controller
     //Manage listings
     public function manage()
     {
-        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]); //
     }
 }
